requirements.txt
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.1
requests>=2.32.0
yfinance>=0.2.40
discord-webhook>=1.3.1
tenacity>=8.2.3
pytz>=2024.1
alpaca-trade-api>=3.1.1
fastapi>=0.111.0
uvicorn>=0.30.0
